+++ projects/swimlane/ngx-ui/src/lib/components/filter/filter.component.ts
@@ -20,7 +20,7 @@ import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import type { AfterViewInit, ComponentRef, QueryList } from '@angular/core';
 
 import { Appearance } from '../../mixins/appearance/appearance.enum';
-import { InViewportMetadata } from 'ng-in-viewport';
+import { InViewportActionEvent } from '../../directives/in-viewport/in-viewport.types';
 import { take } from 'rxjs/operators';
 import { KeyboardKeys } from '../../enums/keyboard-keys.enum';
 import { SelectDropdownOption } from '../select/select-dropdown-option.interface';
@@ -565,13 +565,9 @@ export class FilterComponent implements ControlValueAccessor, AfterViewInit, OnD
     });
   }
 
-  private adjustMenuDirection(event: {
-    [InViewportMetadata]: { entry: IntersectionObserverEntry };
-    target: HTMLElement;
-    visible: boolean;
-  }): void {
-    const { entry } = event[InViewportMetadata];
-    if (!this.forceDownwardOpening && this.isIntersectingBottom(entry) && !this.isIntersectingTop(entry)) {
+  private adjustMenuDirection(event: InViewportActionEvent): void {
+    const entry = event.entry;
+    if (entry && !this.forceDownwardOpening && this.isIntersectingBottom(entry) && !this.isIntersectingTop(entry)) {
       if (this.multiDimension) {
         this._renderer.addClass(this.multiDimensionDropdown.element, 'ngx-multi-dimension-selection--upwards');
       } else {