+++ projects/swimlane/ngx-ui/src/lib/components/multi-dimension-selection/multi-dimension-selection.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { InViewportModule } from 'ng-in-viewport';
+import { InViewportModule } from '../../directives/in-viewport/in-viewport.module';
 
 import { InputModule } from '../input/input.module';
 import { MultiDimensionSelectionComponent } from './multi-dimension-selection.component';