+++ projects/swimlane/ngx-ui/src/public_api.ts
@@ -45,6 +45,9 @@ export * from './lib/directives/validators/pattern-validator/pattern-validator.m
 export * from './lib/directives/validators/pattern-validator/pattern-validator.directive';
 export * from './lib/directives/visibility/visibility.module';
 export * from './lib/directives/visibility/visibility.directive';
+export * from './lib/directives/in-viewport/in-viewport.module';
+export * from './lib/directives/in-viewport/in-viewport.directive';
+export * from './lib/directives/in-viewport/in-viewport.types';
 export * from './lib/directives/autosize-input/autosize-input.module';
 export * from './lib/directives/autosize-input/autosize-input.directive';